# public/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "FastBuild API v4.0",
		"version": "4.0.0",
		"description": "FastBuild 低代码开发平台 REST API\n\n基于 Linus Torvalds 的\"好品味\"设计哲学：**动态表生成 + 智能视图系统 + 四层前缀分离架构**\n\n## 核心设计原则\n\n- **动态表生成**：根据元数据自动创建真实的数据库表\n- **智能视图系统**：每个表自动生成默认视图，支持复杂查询定制\n- **性能优先**：利用 PostgreSQL 的全部能力（索引、约束、事务）\n- **渐进复杂性**：从简单默认视图到复杂物化视图的平滑升级\n- **四层前缀分离架构**：通过URL前缀明确区分API类型和职责",
		"contact": {
			"name": "FastBuild Team",
			"email": "api@fastbuild.com"
		},
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"servers": [
		{
			"url": "https://api.fastbuild.com",
			"description": "Production server"
		},
		{
			"url": "https://preview-api.fastbuild.com",
			"description": "Preview environment"
		},
		{
			"url": "http://localhost:3000",
			"description": "Local development server"
		}
	],
	"paths": {
...
		"/meta/projects/{id}": {
			"get": {
				"operationId": "get-meta-projects-{id}",
				"summary": "获取单个项目",
				"description": "",
				"tags": ["Meta"],
				"parameters": [],
				"responses": {
					"400": {
						"$ref": "#/components/responses/400"
					},
					"401": {
						"$ref": "#/components/responses/401"
					},
					"403": {
						"$ref": "#/components/responses/403"
					},
					"404": {
						"$ref": "#/components/responses/404"
					},
					"500": {
						"$ref": "#/components/responses/500"
					}
				}
			},
			"put": {
				"operationId": "put-meta-projects-{id}",
				"summary": "更新项目",
				"description": "",
				"tags": ["Meta"],
				"parameters": [],
				"responses": {
					"400": {
						"$ref": "#/components/responses/400"
					},
					"401": {
						"$ref": "#/components/responses/401"
					},
					"403": {
						"$ref": "#/components/responses/403"
					},
					"404": {
						"$ref": "#/components/responses/404"
					},
					"500": {
						"$ref": "#/components/responses/500"
					}
				}
			},
			"delete": {
				"operationId": "delete-meta-projects-{id}",
				"summary": "删除项目",
				"description": "",
				"tags": ["Meta"],
				"parameters": [],
				"responses": {
					"400": {
						"$ref": "#/components/responses/400"
					},
					"401": {
						"$ref": "#/components/responses/401"
					},
					"403": {
						"$ref": "#/components/responses/403"
					},
					"404": {
						"$ref": "#/components/responses/404"
					},
					"500": {
						"$ref": "#/components/responses/500"
					}
				}
			}
		},
...
		"/meta/projects": {
			"get": {
				"operationId": "get-meta-projects",
				"summary": "获取项目列表",
				"description": "获取当前用户有权限访问的项目列表，支持分页、搜索和排序",
				"tags": ["s Projects"],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"schema": {
							"type": "string"
						},
						"required": false
					},
					{
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "string"
						},
						"required": false
					},
					{
						"in": "query",
						"name": "search",
						"schema": {
							"type": "string"
						},
						"required": false
					},
					{
						"in": "query",
						"name": "visibility",
						"schema": {
							"type": "string",
							"enum": ["PUBLIC", "PRIVATE"]
						},
						"required": false
					},
					{
						"in": "query",
						"name": "sort",
						"schema": {
							"type": "string",
							"enum": ["name", "createdAt", "updatedAt"]
						},
						"required": false
					},
					{
						"in": "query",
						"name": "order",
						"schema": {
							"type": "string",
							"enum": ["asc", "desc"]
						},
						"required": false
					}
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectListResponseSchema:成功返回项目列表"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/400"
					},
					"401": {
						"$ref": "#/components/responses/401"
					},
					"403": {
						"$ref": "#/components/responses/403"
					},
					"404": {
						"$ref": "#/components/responses/404"
					},
					"500": {
						"$ref": "#/components/responses/500"
					}
				}
			},
			"post": {
				"operationId": "post-meta-projects",
				"summary": "创建新项目",
				"description": "创建一个新的项目，当前用户自动成为项目所有者",
				"tags": ["s Projects"],
				"parameters": [],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProjectSchema"
							}
						}
					},
					"description": "项目创建信息，包含名称、描述和可见性设置"
				},
				"responses": {
					"201": {
						"description": "Successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectResponseSchema:项目创建成功"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/400"
					},
					"401": {
						"$ref": "#/components/responses/401"
					},
					"403": {
						"$ref": "#/components/responses/403"
					},
					"404": {
						"$ref": "#/components/responses/404"
					},
					"500": {
						"$ref": "#/components/responses/500"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"CreateProjectSchema": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 100,
						"pattern": "^[a-zA-Z0-9\\u4e00-\\u9fa5\\s\\-_]+$"
					},
					"slug": {
						"type": "string",
						"minLength": 1,
						"maxLength": 50,
						"pattern": "^[a-z0-9\\-_]+$"
					},
					"description": {
						"type": "string",
						"maxLength": 500,
						"nullable": true
					},
					"visibility": {
						"type": "object"
					}
				},
				"required": ["name", "slug", "visibility", "visibility"]
			},
...
		},
		"responses": {
			"400": {
				"description": "Bad Request",
				"content": {
					"application/json": {
						"schema": {
							"type": "object",
							"properties": {
								"error": {
									"type": "object",
									"properties": {
										"code": {
											"type": "string",
											"description": "Error code",
											"example": "VALIDATION_ERROR"
										},
										"message": {
											"type": "string",
											"description": "Human readable error message",
											"example": "Invalid input data"
										},
										"details": {
											"oneOf": [
												{
													"type": "array",
													"items": {
														"type": "string"
													},
													"description": "Array of error details"
												},
												{
													"type": "object",
													"description": "Detailed error information"
												}
											]
										}
									},
									"required": ["code", "message"]
								}
							}
						}
					}
				}
			},
			"401": {
				"description": "Unauthorized",
				"content": {
					"application/json": {
						"schema": {
							"type": "object",
							"properties": {
								"error": {
									"type": "object",
									"properties": {
										"code": {
											"type": "string",
											"description": "Error code",
											"example": "VALIDATION_ERROR"
										},
										"message": {
											"type": "string",
											"description": "Human readable error message",
											"example": "Invalid input data"
										},
										"details": {
											"oneOf": [
												{
													"type": "array",
													"items": {
														"type": "string"
													},
													"description": "Array of error details"
												},
												{
													"type": "object",
													"description": "Detailed error information"
												}
											]
										}
									},
									"required": ["code", "message"]
								}
							}
						}
					}
				}
			},
			"403": {
				"description": "Forbidden",
				"content": {
					"application/json": {
						"schema": {
							"type": "object",
							"properties": {
								"error": {
									"type": "object",
									"properties": {
										"code": {
											"type": "string",
											"description": "Error code",
											"example": "VALIDATION_ERROR"
										},
										"message": {
											"type": "string",
											"description": "Human readable error message",
											"example": "Invalid input data"
										},
										"details": {
											"oneOf": [
												{
													"type": "array",
													"items": {
														"type": "string"
													},
													"description": "Array of error details"
												},
												{
													"type": "object",
													"description": "Detailed error information"
												}
											]
										}
									},
									"required": ["code", "message"]
								}
							}
						}
					}
				}
			},
			"404": {
				"description": "Not Found",
				"content": {
					"application/json": {
						"schema": {
							"type": "object",
							"properties": {
								"error": {
									"type": "object",
									"properties": {
										"code": {
											"type": "string",
											"description": "Error code",
											"example": "VALIDATION_ERROR"
										},
										"message": {
											"type": "string",
											"description": "Human readable error message",
											"example": "Invalid input data"
										},
										"details": {
											"oneOf": [
												{
													"type": "array",
													"items": {
														"type": "string"
													},
													"description": "Array of error details"
												},
												{
													"type": "object",
													"description": "Detailed error information"
												}
											]
										}
									},
									"required": ["code", "message"]
								}
							}
						}
					}
				}
			},
...
			"500": {
				"description": "Internal Server Error",
				"content": {
					"application/json": {
						"schema": {
							"type": "object",
							"properties": {
								"error": {
									"type": "object",
									"properties": {
										"code": {
											"type": "string",
											"description": "Error code",
											"example": "VALIDATION_ERROR"
										},
										"message": {
											"type": "string",
											"description": "Human readable error message",
											"example": "Invalid input data"
										},
										"details": {
											"oneOf": [
												{
													"type": "array",
													"items": {
														"type": "string"
													},
													"description": "Array of error details"
												},
												{
													"type": "object",
													"description": "Detailed error information"
												}
											]
										}
									},
									"required": ["code", "message"]
								}
							}
						}
					}
				}
			}
		}
	}
}